requirements.txt

pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

lightgbm>=4.0.0
networkx>=3.1
node2vec>=0.4.6
gensim>=4.3.0

shap>=0.43.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
httpx>=0.25.0

streamlit>=1.28.0
plotly>=5.17.0
pyvis>=0.3.2

kafka-python>=2.0.2
confluent-kafka>=2.2.0

mlflow>=2.8.0

python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0.1
click>=8.1.7
loguru>=0.7.2
prometheus-client>=0.18.0

pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
